idam-ffi: move message encryption to separate aranya-msg-ffi module
- Remove unused compute_change_id function
- Create aranya-msg-ffi for message encryption and group key management
- Keep aranya-idam-ffi focused on identity and access management

Closes #271
